added pages to view by category: 1) all books only 2) all games only 3) all games only

+++ src/main/java/com/mindteck/booksandstuff/service/impl/RoleServiceImpl.java
@@ -27,7 +27,7 @@ public class RoleServiceImpl implements RoleService {
 		List<Role> roles = roleDao.getRoles();
 		Map<String, String> roleMap = new HashMap<String,String>();
 		for(Role role:roles){
-			roleMap.put(Integer.toString(role.getId()), role.getName());
+			roleMap.put(Long.toString(role.getId()), role.getName());
 			System.out.println(role.getName()+" / "+role.getId());
 		};
 		return roleMap;
